app/requirements.txt
redis[hiredis]==4.5.2
fastapi==0.95.0
orjson==3.8.7
pydantic==1.10.6
uvicorn==0.21.1
uvloop==0.17.0
aiohttp==3.8.4
pyjwt==2.6.0
